Move rest tests into dedicated package. Fix unused imports in app module. Only run rest tests from failsafe for now.
