+++ tests/acceptance-test/src/test/java/com/quorum/tessera/test/grpc/SendGrpcIT.java
@@ -8,8 +8,10 @@ import io.grpc.ManagedChannelBuilder;
 import static org.assertj.core.api.Assertions.assertThat;
 import org.junit.After;
 import org.junit.Before;
+import org.junit.Ignore;
 import org.junit.Test;
 
+@Ignore
 public class SendGrpcIT {
 
     private ManagedChannel channel;
@@ -18,7 +20,7 @@ public class SendGrpcIT {
 
     @Before
     public void onSetUp() {
-        channel = ManagedChannelBuilder.forAddress("127.0.0.1", 9990)
+        channel = ManagedChannelBuilder.forAddress("127.0.0.1", 8080)
                 .usePlaintext()
                 .build();
