+++ tests/acceptance-test/src/test/java/com/quorum/tessera/test/rest/ResendAllIT.java
@@ -1,4 +1,4 @@
-package com.quorum.tessera.test;
+package com.quorum.tessera.test.rest;
 
 import com.quorum.tessera.api.model.ResendRequest;
 import com.quorum.tessera.api.model.ResendRequestType;
