+++ tests/acceptance-test/src/test/java/com/quorum/tessera/test/rest/SendIT.java
@@ -1,4 +1,4 @@
-package com.quorum.tessera.test;
+package com.quorum.tessera.test.rest;
 
 import org.junit.Test;
 import org.slf4j.Logger;
