+++ tests/acceptance-test/src/test/resources/config3.json
@@ -7,7 +7,6 @@
     },
     "server": {
         "port": 8082,
-        "grpcPort":9992,
         "hostName": "http://localhost"
     },
     "peer": [
